compiler: Record data alloc cleanups
- Remove GenSizeOf. We already calculate the precise size of the record
  data while generating types.

- Consistently use 32bit sizes when allocating record data. We already
  used 32bit sizes in the class map and during cloning meaning 64bit
  sizes were not properly supported.

+++ runtime/core/recorddata.cpp
@@ -8,7 +8,7 @@ extern "C"
 
 using namespace lliby;
 
-void *llcore_record_data_alloc(std::uint64_t size)
+void *llcore_record_data_alloc(std::uint32_t size)
 {
 	return RecordLikeCell::allocateRecordData(size);
 }
